Hardcode visitor iterators per bucket to 1
We have an SPI iterate() invariant that states the provider shall never
observe concurrent calls for the same iterator ID. If GetIter operations
can operate in shared locking mode, this invariant will no longer hold
if multiple GetIters for a single iterator ID can be pipelined in the
queue. We therefore ignore the config entirely.

This is not expected to cause any performance regressions in practice.